Fix issue that padlock icon is not animating while expanding QS
In case that some of settings are disallowed with device policy, padlock
icon appears at the beside QSTile's label. The issue is that padlock
icon is not animated with QSAnimator while expanding or collapsing
QSPanel. This is because QSAnimator makes firstPageAnimator without
padlock. So, we should apply animator not to label but to label's parent
in order to animate padlock together.

Bug:29588337
Change-Id: I85d85e729ca9d86af9cb8fcdc7af2a909ce6bb87